Reorganisation du code
- Split des vues (`contents`, `publications` et `published`)
- Split des modeles (`database` et `versioned`)
- Migration des fonctions statiques et exceptions dans `utils.py`
- Suppression de `zds/utils/tutorialv2.py`
- Creation d'URLs de valiation (`validation:*`)
- Corrections en consequences des templates et imports
