+++ doc/source/back-end-code/tutorialv2.rst
@@ -6,16 +6,42 @@ Module situé dans ``zds/tutorialv2/``.
 
 .. contents:: Fichiers documentés :
 
-Modèles (``models.py``)
-=======================
+Modèles (``models/``)
+=====================
 
-.. automodule:: zds.tutorialv2.models
+Modèles de la base de donnée (``models_database.py``)
+-----------------------------------------------------
+
+.. automodule:: zds.tutorialv2.models_database
     :members:
 
-Vues (``views.py``)
-===================
 
-.. automodule:: zds.tutorialv2.views
+Modèles "versionnés" (``models_versioned.py``)
+----------------------------------------------
+
+.. automodule:: zds.tutorialv2.models.models_versioned
+    :members:
+
+Vues (``views/``)
+=================
+
+Contenus (``views/views_contents.py``)
+--------------------------------------
+
+.. automodule:: zds.tutorialv2.views.views_contents
+    :members:
+
+
+Validations (``views/views_validations.py``)
+--------------------------------------------
+
+.. automodule:: zds.tutorialv2.views_validations
+    :members:
+
+Contenus publiés (``views/views_published.py``)
+-----------------------------------------------
+
+.. automodule:: zds.tutorialv2.views_published
     :members:
 
 Mixins (``mixins.py``)
