+++ templates/tutorialv2/validation/index.html
@@ -78,7 +78,7 @@
                                             {% if not forloop.first %}
                                                 -
                                             {% endif %}
-                                            <a href="{% url "content:list-validation" %}?subcategory={{ subcategory.pk }}">
+                                            <a href="{% url "validation:list" %}?subcategory={{ subcategory.pk }}">
                                                 {{ subcategory.title }}
                                             </a>
                                         {% endfor %}
@@ -88,9 +88,9 @@
                                 </td>
                                 <td>
                                     {% if validation.content.type == "ARTICLE" %}
-                                        <a href="{% url "content:list-validation" %}?type=article">Article</a>
+                                        <a href="{% url "validation:list" %}?type=article">Article</a>
                                     {% else %}
-                                        <a href="{% url "content:list-validation" %}?type=tuto">Tutoriel</a>
+                                        <a href="{% url "validation:list" %}?type=tuto">Tutoriel</a>
                                     {% endif %}
                                 </td>
                                 <td>
@@ -103,7 +103,7 @@
                                 </td>
                                 <td>
                                     {% captureas reservation_url %}
-                                        {% url "content:reserve-validation" validation.pk %}
+                                        {% url "validation:reserve" validation.pk %}
                                     {% endcaptureas %}
                                     {% include "misc/validation.part.html" %}
                                 </td>
@@ -131,28 +131,28 @@
         <h3>{% trans "Filtres" %}</h3>
         <ul>
             <li>
-                <a href="{% url "content:list-validation" %}?type=reserved" class="ico-after tick green {% if request.GET.type == "reserved" %}selected{% endif %}">
+                <a href="{% url "validation:list" %}?type=reserved" class="ico-after tick green {% if request.GET.type == "reserved" %}selected{% endif %}">
                     {% trans "En cours de validation" %}
                 </a>
             </li>
             <li>
-                <a href="{% url "content:list-validation" %}?type=orphan" class="ico-after tick {% if request.GET.type == "orphan" %}selected{% endif %}">
+                <a href="{% url "validation:list" %}?type=orphan" class="ico-after tick {% if request.GET.type == "orphan" %}selected{% endif %}">
                     {% trans "En attente de validateur" %}
                 </a>
             </li>
             <li>
-                <a href="{% url "content:list-validation" %}?type=article" class="ico-after view {% if request.GET.type == "article" %}selected{% endif %}">
+                <a href="{% url "validation:list" %}?type=article" class="ico-after view {% if request.GET.type == "article" %}selected{% endif %}">
                     {% trans "Articles" %}
                 </a>
             </li>
             <li>
-                <a href="{% url "content:list-validation" %}?type=tuto" class="ico-after view {% if request.GET.type == "tuto" %}selected{% endif %}">
+                <a href="{% url "validation:list" %}?type=tuto" class="ico-after view {% if request.GET.type == "tuto" %}selected{% endif %}">
                     {% trans "Tutoriels" %}
                 </a>
             </li>
             {% if request.GET.type or request.GET.subcategory %}
                 <li>
-                    <a href="{% url "content:list-validation" %}" class="ico-after cross red">
+                    <a href="{% url "validation:list" %}" class="ico-after cross red">
                         {% trans "Annuler le filtre" %}
                     </a>
                 </li>
