+++ templates/tutorialv2/view/content_online.html
@@ -107,7 +107,7 @@
                     </a>
                 </li>
                 <li>
-                    <a href="{% url "content:history-validation" content.pk content.slug %}" class="ico-after history blue">
+                    <a href="{% url "validation:history" content.pk content.slug %}" class="ico-after history blue">
                         {% trans "Historique validation" %}
                     </a>
                 </li>
