+++ zds/tutorialv2/utils.py
@@ -1,6 +1,8 @@
 # coding: utf-8
+from collections import OrderedDict
 
 import shutil
+from git import Repo, Actor
 import os
 from datetime import datetime
 import copy
@@ -10,19 +12,11 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.template.loader import render_to_string
 from django.utils.translation import ugettext_lazy as _
 
-from zds.tutorialv2.models import PublishableContent, ContentRead, Container, Extract, PublishedContent
 from zds import settings
-from zds.utils import get_current_user
+from zds.settings import ZDS_APP
+from zds.utils import get_current_user, slugify
+from zds.utils.models import Licence
 from zds.utils.templatetags.emarkdown import emarkdown
-
-
-class TooDeepContainerError(ValueError):
-    """
-    Exception used to represent the fact you can't add a container to a level greater than two
-    """
-
-    def __init__(self, *args, **kwargs):
-        super(TooDeepContainerError, self).__init__(*args, **kwargs)
 
 
 def search_container_or_404(base_content, kwargs_array):
@@ -33,6 +27,9 @@ def search_container_or_404(base_content, kwargs_array):
     :return: the Container object we were searching for
     :raise Http404 if no suitable container is found
     """
+
+    from zds.tutorialv2.models.models_versioned import Container
+
     container = None
     if isinstance(kwargs_array, basestring):
         dic = {}
@@ -77,6 +74,9 @@ def search_extract_or_404(base_content, kwargs_array):
     :return: the Extract object
     :raise: Http404 if not found
     """
+
+    from zds.tutorialv2.models.models_versioned import Extract
+
     # if the extract is at a depth of 3 we get the first parent container
     container = search_container_or_404(base_content, kwargs_array)
 
@@ -92,40 +92,15 @@ def search_extract_or_404(base_content, kwargs_array):
     return extract
 
 
-def get_last_tutorials():
-    """
-    :return: last issued tutorials
-    """
-    n = settings.ZDS_APP['tutorial']['home_number']
-    tutorials = PublishableContent.objects.all()\
-        .exclude(type="ARTICLE")\
-        .exclude(sha_public__isnull=True)\
-        .exclude(sha_public__exact='')\
-        .order_by('-pubdate')[:n]
-
-    return tutorials
-
-
-def get_last_articles():
-    """
-    :return: last issued articles
-    """
-    n = settings.ZDS_APP['tutorial']['home_number']
-    articles = PublishableContent.objects.all()\
-        .exclude(type="TUTO")\
-        .exclude(sha_public__isnull=True)\
-        .exclude(sha_public__exact='')\
-        .order_by('-pubdate')[:n]
-
-    return articles
-
-
 def never_read(content, user=None):
     """Check if a content note feed has been read by an user since its last post was added.
 
     :param content: the content to check
     :return: `True` if it is the case, `False` otherwise
     """
+
+    from zds.tutorialv2.models.models_database import ContentRead
+
     if user is None:
         user = get_current_user()
 
@@ -139,6 +114,9 @@ def mark_read(content):
 
     :param content: the content to mark
     """
+
+    from zds.tutorialv2.models.models_database import ContentRead
+
     if content.last_note is not None:
         ContentRead.objects.filter(
             content=content,
@@ -148,6 +126,15 @@ def mark_read(content):
             content=content,
             user=get_current_user())
         a.save()
+
+
+class TooDeepContainerError(ValueError):
+    """
+    Exception used to represent the fact you can't add a container to a level greater than two
+    """
+
+    def __init__(self, *args, **kwargs):
+        super(TooDeepContainerError, self).__init__(*args, **kwargs)
 
 
 def try_adopt_new_child(adoptive_parent, child):
@@ -160,6 +147,8 @@ def try_adopt_new_child(adoptive_parent, child):
     :raise TypeError: if the adoptive parent is not allowed to adopt the child due to its type
     :raise TooDeepContainerError: if the child is a container that is too deep to be adopted by the proposed parent
     """
+
+    from zds.tutorialv2.models.models_versioned import Container, Extract
 
     if isinstance(child, Extract):
         if not adoptive_parent.can_add_extract():
@@ -180,6 +169,9 @@ def get_target_tagged_tree(movable_child, root):
     :return: an array of tuples that represent the capacity of movable_child to be moved near another child
     check get_target_tagged_tree_for_extract and get_target_tagged_tree_for_container for format
     """
+
+    from zds.tutorialv2.models.models_versioned import Extract
+
     if isinstance(movable_child, Extract):
         return get_target_tagged_tree_for_extract(movable_child, root)
     else:
@@ -194,6 +186,9 @@ def get_target_tagged_tree_for_extract(movable_child, root):
     :return: an array of tuples that represent the capacity of movable_child to be moved near another child
     tuples are ``(relative_path, title, level, can_be_a_target)``
     """
+
+    from zds.tutorialv2.models.models_versioned import Extract
+
     target_tagged_tree = []
     for child in root.traverse(False):
         if isinstance(child, Extract):
@@ -258,6 +253,11 @@ def publish_container(db_object, base_dir, container):
     :type container: Container
     :raise FailureDuringPublication: if anything goes wrong
     """
+
+    from zds.tutorialv2.models.models_versioned import Container
+
+    if not isinstance(container, Container):
+        raise FailureDuringPublication(_(u'Le conteneur n\'en est pas un !'))
 
     # TODO: images stuff !!
 
@@ -344,8 +344,10 @@ def publish_content(db_object, versioned, is_major_update=True):
     :type is_major_update: bool
     :raise FailureDuringPublication: if something goes wrong
     :return: the published representation
-    :rtype: PublishedContent
+    :rtype: zds.tutorialv2.models.models_database.PublishedContent
     """
+
+    from zds.tutorialv2.models.models_database import PublishedContent
 
     if is_major_update:
         versioned.pubdate = datetime.now()
@@ -437,6 +439,8 @@ def unpublish_content(db_object):
     :return; `True` if unpublished, `False otherwise`
     """
 
+    from zds.tutorialv2.models.models_database import PublishedContent
+
     try:
         public_version = PublishedContent.objects.get(content=db_object)
 
@@ -458,3 +462,306 @@ def unpublish_content(db_object):
         pass
 
     return False
+
+
+class BadManifestError(Exception):
+    """ The exception that is raised when the manifest.json contains errors """
+
+    def __init__(self, *args, **kwargs):
+        super(BadManifestError, self).__init__(*args, **kwargs)
+
+
+def get_content_from_json(json, sha, slug_last_draft, public=False):
+    """
+    Transform the JSON formated data into `VersionedContent`
+    :param json: JSON data from a `manifest.json` file
+    :param sha: version
+    :param public: the function will fill a PublicContent instead of a VersionedContent if `True`
+    :return: a Public/VersionedContent with all the information retrieved from JSON
+    """
+
+    from zds.tutorialv2.models.models_versioned import Container, Extract, VersionedContent, PublicContent
+
+    if 'version' in json and json['version'] == 2:
+        # create and fill the container
+        if not public:
+            versioned = VersionedContent(sha, 'TUTORIAL', json['title'], json['slug'], slug_last_draft)
+        else:
+            versioned = PublicContent(sha, 'TUTORIAL', json['title'], json['slug'])
+
+        # fill metadata :
+        if 'description' in json:
+            versioned.description = json['description']
+
+        if 'type' in json:
+            if json['type'] == 'ARTICLE' or json['type'] == 'TUTORIAL':
+                versioned.type = json['type']
+
+        if 'licence' in json:
+            versioned.licence = Licence.objects.filter(code=json['licence']).first()
+        else:
+            versioned.licence = Licence.objects.filter(pk=settings.ZDS_APP['content']['default_license_pk']).first()
+
+        if 'introduction' in json:
+            versioned.introduction = json['introduction']
+        if 'conclusion' in json:
+            versioned.conclusion = json['conclusion']
+
+        # then, fill container with children
+        fill_containers_from_json(json, versioned)
+    else:
+        # MINIMUM (!) fallback for version 1.0
+        if "type" in json:
+            if json['type'] == 'article':
+                _type = 'ARTICLE'
+            else:
+                _type = "TUTORIAL"
+        else:
+            _type = "ARTICLE"
+
+        if not public:
+            versioned = VersionedContent(sha, _type, json['title'], slug_last_draft)
+        else:
+            versioned = PublicContent(sha, _type, json['title'], slug_last_draft)
+
+        if 'description' in json:
+            versioned.description = json['description']
+        if "introduction" in json:
+            versioned.introduction = json["introduction"]
+        if "conclusion" in json:
+            versioned.conclusion = json["conclusion"]
+        if 'licence' in json:
+            versioned.licence = Licence.objects.filter(code=json['licence']).first()
+        else:
+            versioned.licence = Licence.objects.filter(pk=settings.ZDS_APP['content']['default_license_pk']).first()
+
+        if _type == 'ARTICLE':
+            extract = Extract(json['title'], '')
+            if 'text' in json:
+                extract.text = json['text']  # probably "text.md" !
+            versioned.add_extract(extract, generate_slug=True)
+
+        else:  # it's a tutorial
+            if json['type'] == 'MINI' and 'chapter' in json and 'extracts' in json['chapter']:
+                for extract in json['chapter']['extracts']:
+                    new_extract = Extract(extract['title'], '{}_{}'.format(extract['pk'], slugify(extract['title'])))
+                    if 'text' in extract:
+                        new_extract.text = extract['text']
+                    versioned.add_extract(new_extract, generate_slug=False)
+
+            elif json['type'] == 'BIG' and 'parts' in json:
+                for part in json['parts']:
+                    new_part = Container(part['title'], '{}_{}'.format(part['pk'], slugify(part['title'])))
+                    if 'introduction' in part:
+                        new_part.introduction = part['introduction']
+                    if 'conclusion' in part:
+                        new_part.conclusion = part['conclusion']
+                    versioned.add_container(new_part, generate_slug=False)
+
+                    if 'chapters' in part:
+                        for chapter in part['chapters']:
+                            new_chapter = Container(
+                                chapter['title'], '{}_{}'.format(chapter['pk'], slugify(chapter['title'])))
+                            if 'introduction' in chapter:
+                                new_chapter.introduction = chapter['introduction']
+                            if 'conclusion' in chapter:
+                                new_chapter.conclusion = chapter['conclusion']
+                            new_part.add_container(new_chapter, generate_slug=False)
+
+                            if 'extracts' in chapter:
+                                for extract in chapter['extracts']:
+                                    new_extract = Extract(
+                                        extract['title'], '{}_{}'.format(extract['pk'], slugify(extract['title'])))
+                                    if 'text' in extract:
+                                        new_extract.text = extract['text']
+                                    new_chapter.add_extract(new_extract, generate_slug=False)
+
+    return versioned
+
+
+def fill_containers_from_json(json_sub, parent):
+    """Function which call itself to fill container
+
+    :param json_sub: dictionary from "manifest.json"
+    :param parent: the container to fill
+    """
+
+    from zds.tutorialv2.models.models_versioned import Container, Extract
+
+    if 'children' in json_sub:
+        for child in json_sub['children']:
+            if child['object'] == 'container':
+                slug = ''
+                try:
+                    slug = child['slug']
+                except KeyError:
+                    pass
+                new_container = Container(child['title'], slug)
+                if 'introduction' in child:
+                    new_container.introduction = child['introduction']
+                if 'conclusion' in child:
+                    new_container.conclusion = child['conclusion']
+                parent.add_container(new_container, generate_slug=(slug != ''))
+                if 'children' in child:
+                    fill_containers_from_json(child, new_container)
+            elif child['object'] == 'extract':
+                slug = ''
+                try:
+                    slug = child['slug']
+                except KeyError:
+                    pass
+                new_extract = Extract(child['title'], slug)
+
+                if 'text' in child:
+                    new_extract.text = child['text']
+
+                parent.add_extract(new_extract, generate_slug=(slug != ''))
+            else:
+                raise BadManifestError(_(u'Type d\'objet inconnu : {}').format(child['object']))
+
+
+def init_new_repo(db_object, introduction_text, conclusion_text, commit_message='', do_commit=True):
+    """Create a new repository in `settings.ZDS_APP['contents']['private_repo']` to store the files for a new content.
+    Note that `db_object.sha_draft` will be set to the good value
+
+    :param db_object: `PublishableContent` (WARNING: should have a valid `slug`, so previously saved)
+    :param introduction_text: introduction from form
+    :param conclusion_text: conclusion from form
+    :param commit_message : set a commit message instead of the default one
+    :param do_commit: do commit if `True`
+    :return: `VersionedContent` object
+    """
+
+    from zds.tutorialv2.models.models_versioned import VersionedContent
+
+    # create directory
+    path = db_object.get_repo_path()
+    if not os.path.isdir(path):
+        os.makedirs(path, mode=0o777)
+
+    # init repo:
+    Repo.init(path, bare=False)
+
+    # create object
+    versioned_content = VersionedContent(None, db_object.type, db_object.title, db_object.slug)
+
+    # fill some information that are missing :
+    versioned_content.licence = db_object.licence
+    versioned_content.description = db_object.description
+
+    # perform changes:
+    if commit_message == '':
+        commit_message = u'Création du contenu'
+
+    sha = versioned_content.repo_update(
+        db_object.title, introduction_text, conclusion_text, commit_message=commit_message, do_commit=do_commit)
+
+    # update sha:
+    if do_commit:
+        db_object.sha_draft = sha
+        db_object.sha_beta = None
+        db_object.sha_public = None
+        db_object.sha_validation = None
+
+        db_object.save()
+
+    return versioned_content
+
+
+def get_commit_author():
+    """
+    :return: correctly formatted commit author for `repo.index.commit()`
+    """
+    user = get_current_user()
+
+    if user:
+        aut_user = str(user.pk)
+        aut_email = None
+
+        if hasattr(user, 'email'):
+            aut_email = user.email
+
+    else:
+        aut_user = ZDS_APP['member']['bot_account']
+        aut_email = None
+
+    if aut_email is None or aut_email.strip() == "":
+        aut_email = "inconnu@{}".format(settings.ZDS_APP['site']['dns'])
+    return {'author': Actor(aut_user, aut_email), 'committer': Actor(aut_user, aut_email)}
+
+
+def export_extract(extract):
+    """
+    Export an extract to a dictionary
+    :param extract: extract to export
+    :return: dictionary containing the information
+    """
+    dct = OrderedDict()
+    dct['object'] = 'extract'
+    dct['slug'] = extract.slug
+    dct['title'] = extract.title
+
+    if extract.text:
+        dct['text'] = extract.text
+
+    return dct
+
+
+def export_container(container):
+    """
+    Export a container to a dictionary
+    :param container: the container
+    :return: dictionary containing the information
+    """
+    dct = OrderedDict()
+    dct['object'] = "container"
+    dct['slug'] = container.slug
+    dct['title'] = container.title
+
+    if container.introduction:
+        dct['introduction'] = container.introduction
+
+    if container.conclusion:
+        dct['conclusion'] = container.conclusion
+
+    dct['children'] = []
+
+    if container.has_sub_containers():
+        for child in container.children:
+            dct['children'].append(export_container(child))
+    elif container.has_extracts():
+        for child in container.children:
+            dct['children'].append(export_extract(child))
+
+    return dct
+
+
+def export_content(content):
+    """
+    Export a content to dictionary in order to store them in a JSON file
+    :param content: content to be exported
+    :return: dictionary containing the information
+    """
+    dct = export_container(content)
+
+    # append metadata :
+    dct['version'] = 2  # to recognize old and new version of the content
+    dct['description'] = content.description
+    dct['type'] = content.type
+    if content.licence:
+        dct['licence'] = content.licence.code
+
+    return dct
+
+
+def default_slug_pool():
+    """
+    :return: the forbidden slugs in the edition system
+    :rtype: dict
+    """
+
+    return {'introduction': 1, 'conclusion': 1}  # forbidden slugs
+
+
+class InvalidOperationError(RuntimeError):
+    pass
