+++ zds/tutorialv2/views/views_validations.py
@@ -1,0 +1,486 @@
+# coding: utf-8
+from datetime import datetime
+from django.conf import settings
+from django.contrib import messages
+from django.contrib.auth.models import User
+from django.core.exceptions import PermissionDenied
+from django.core.urlresolvers import reverse
+from django.db.models import Q
+from django.shortcuts import get_object_or_404, redirect
+from django.template.loader import render_to_string
+from django.utils.translation import ugettext_lazy as _
+from django.views.generic import ListView, FormView
+from zds.member.decorator import LoginRequiredMixin, PermissionRequiredMixin, LoggedWithReadWriteHability
+from zds.tutorialv2.forms import AskValidationForm, RejectValidationForm, AcceptValidationForm, RevokeValidationForm
+from zds.tutorialv2.mixins import SingleContentFormViewMixin, SingleContentDetailViewMixin, ModalFormView
+from zds.tutorialv2.models.models_database import Validation, PublishableContent
+from zds.tutorialv2.utils import publish_content, FailureDuringPublication, unpublish_content
+from zds.utils.models import SubCategory
+from zds.utils.mps import send_mp
+
+
+class ValidationListView(LoginRequiredMixin, PermissionRequiredMixin, ListView):
+    """List the validations, with possibilities of filters"""
+
+    permissions = ["tutorial.change_tutorial"]
+    context_object_name = "validations"
+    template_name = "tutorialv2/validation/index.html"
+
+    def get_queryset(self):
+
+        # TODO: many filter at the same time ?
+        # TODO: paginate ?
+
+        queryset = Validation.objects\
+            .prefetch_related("validator")\
+            .prefetch_related("content")\
+            .prefetch_related("content__authors")\
+            .filter(Q(status="PENDING") | Q(status="PENDING_V"))
+
+        # filtering by type
+        try:
+            type_ = self.request.GET["type"]
+            if type_ == "orphan":
+                queryset = queryset.filter(
+                    validator__isnull=True,
+                    status="PENDING")
+            if type_ == "reserved":
+                queryset = queryset.filter(
+                    validator__isnull=True,
+                    status="PENDING_V")
+            if type_ == "article":
+                queryset = queryset.filter(
+                    content__type="ARTICLE")
+            if type_ == "tuto":
+                queryset = queryset.filter(
+                    content__type="TUTORIAL")
+            else:
+                raise KeyError()
+        except KeyError:
+            pass
+
+        # filtering by category
+        try:
+            category = get_object_or_404(SubCategory, pk=self.request.GET["subcategory"])
+            queryset = queryset.filter(content__subcategory__in=[category])
+        except KeyError:
+            pass
+
+        return queryset.order_by("date_proposition").all()
+
+    def get_context_data(self, **kwargs):
+        context = super(ValidationListView, self).get_context_data(**kwargs)
+
+        if 'subcategory' in self.request.GET:
+            context['category'] = get_object_or_404(SubCategory, pk=self.request.GET["subcategory"])
+            # TODO : two times the same request, here
+
+        return context
+
+
+class AskValidationForContent(LoggedWithReadWriteHability, SingleContentFormViewMixin):
+    """User ask validation for his tutorial. Staff member can also to that"""
+
+    prefetch_all = False
+    form_class = AskValidationForm
+    must_be_author = True
+    authorized_for_staff = True  # an admin could ask validation for a content
+    only_draft_version = False
+    modal_form = True
+
+    def dispatch(self, *args, **kwargs):
+        if "version" in self.request.POST:
+            self.sha = self.request.POST["version"]
+        return super(AskValidationForContent, self).dispatch(*args, **kwargs)
+
+    def get_form_kwargs(self):
+        kwargs = super(AskValidationForContent, self).get_form_kwargs()
+        kwargs['content'] = self.versioned_object
+        return kwargs
+
+    def form_valid(self, form):
+
+        # test if admin or author
+        """"if not self.request.user in self.object.authors.all() \
+                and not self.request.user.has_perm('tutorial.change_tutorial'):
+            raise PermissionDenied"""
+
+        old_validation = Validation.objects.filter(
+            content__pk=self.object.pk, status__in=['PENDING', 'PENDING_V']).first()
+
+        if old_validation:  # if an old validation exists, cancel it !
+            old_validator = old_validation.validator
+            old_validation.status = 'CANCEL'
+            old_validation.date_validation = datetime.now()
+            old_validation.save()
+        else:
+            old_validator = None
+
+        # create a "validation" object
+        validation = Validation()
+        validation.content = self.object
+        validation.date_proposition = datetime.now()
+        validation.comment_authors = form.cleaned_data['text']
+        validation.version = form.cleaned_data['version']
+        validation.save()
+
+        # warn the former validator that an update has been made, if any
+        if old_validator:
+            validation.validator = old_validator
+            # TODO: why not let the validator be the same using PENDING_V ?
+
+            bot = get_object_or_404(User, username=settings.ZDS_APP['member']['bot_account'])
+            msg = render_to_string(
+                'tutorialv2/messages/validation_change.md',
+                {
+                    'content': self.versioned_object,
+                    'validator': validation.validator.username,
+                    'url': self.versioned_object.get_absolute_url() + '?version=' + form.cleaned_data['version'],
+                    'url_history': reverse('content:history', args=[self.object.pk, self.object.slug])
+                })
+
+            send_mp(
+                bot,
+                [old_validator],
+                _(u"Une nouvelle version a été envoyée en validation"),
+                self.versioned_object.title,
+                msg,
+                False,
+            )
+
+        # update the content with the source and the version of the validation
+        self.object.source = form.cleaned_data["source"]
+        self.object.sha_validation = validation.version
+        self.object.save()
+
+        messages.success(self.request, _(u"Votre demande de validation a été envoyée à l'équipe."))
+
+        self.success_url = self.versioned_object.get_absolute_url(version=self.sha)
+        return super(AskValidationForContent, self).form_valid(form)
+
+
+class CancelValidation(LoginRequiredMixin, FormView):
+    """The user or an admin cancel the validation process"""
+
+    def post(self, request, *args, **kwargs):
+
+        user = self.request.user
+
+        validation = Validation.objects\
+            .filter(pk=self.kwargs['pk'])\
+            .prefetch_related('content')\
+            .prefetch_related('content__authors')\
+            .last()
+
+        if not validation:
+            raise PermissionDenied
+
+        if validation.status not in ['PENDING', 'PENDING_V']:
+            raise PermissionDenied  # cannot cancel a validation that is already accepted or rejected
+
+        if user not in validation.content.authors.all() and not user.has_perms('tutorial.change_tutorial'):
+            raise PermissionDenied
+
+        versioned = validation.content.load_version(sha=validation.version)
+
+        # reject validation:
+        validation.status = "CANCEL"
+        validation.date_validation = datetime.now()
+        validation.save()
+
+        validation.content.sha_validation = None
+        validation.content.save()
+
+        # warn the former validator that the all thing have been canceled
+        if validation.validator:
+            bot = get_object_or_404(User, username=settings.ZDS_APP['member']['bot_account'])
+            msg = render_to_string(
+                'tutorialv2/messages/validation_cancel.md',
+                {
+                    'content': versioned,
+                    'validator': validation.validator.username,
+                    'url': versioned.get_absolute_url() + '?version=' + validation.version
+                })
+
+            send_mp(
+                bot,
+                [validation.validator],
+                _(u"Demande de validation annulée").format(),
+                versioned.title,
+                msg,
+                False,
+            )
+
+        messages.info(self.request, _(u'La validation de ce contenu a bien été annulée'))
+
+        return redirect(
+            reverse("content:view", args=[validation.content.pk, validation.content.slug]) +
+            "?version=" + validation.version)
+
+
+class ReserveValidation(LoginRequiredMixin, PermissionRequiredMixin, FormView):
+    """Reserve or remove the reservation on a content"""
+
+    permissions = ["tutorial.change_tutorial"]
+
+    def post(self, request, *args, **kwargs):
+        validation = get_object_or_404(Validation, pk=kwargs["pk"])
+        if validation.validator:
+            validation.validator = None
+            validation.date_reserve = None
+            validation.status = "PENDING"
+            validation.save()
+            messages.info(request, _(u"Ce contenu n'est plus réservé"))
+            return redirect(reverse("validation:list"))
+        else:
+            validation.validator = request.user
+            validation.date_reserve = datetime.now()
+            validation.status = "PENDING_V"
+            validation.save()
+            messages.info(request, _(u"Ce contenu a bien été réservé par {0}").format(request.user.username))
+
+            return redirect(
+                reverse("content:view", args=[validation.content.pk, validation.content.slug]) +
+                "?version=" + validation.version
+            )
+
+
+class HistoryOfValidationDisplay(LoginRequiredMixin, PermissionRequiredMixin, SingleContentDetailViewMixin):
+
+    model = PublishableContent
+    permissions = ["tutorial.change_tutorial"]
+    template_name = "tutorialv2/validation/history.html"
+
+    def get_context_data(self, **kwargs):
+        context = super(HistoryOfValidationDisplay, self).get_context_data()
+
+        context["validations"] = Validation.objects\
+            .prefetch_related("validator")\
+            .filter(content__pk=self.object.pk)\
+            .order_by("date_proposition").all()
+
+        return context
+
+
+class RejectValidation(LoginRequiredMixin, PermissionRequiredMixin, ModalFormView):
+    """Reject the publication"""
+
+    permissions = ["tutorial.change_tutorial"]
+    form_class = RejectValidationForm
+
+    modal_form = True
+
+    def get_form_kwargs(self):
+        kwargs = super(RejectValidation, self).get_form_kwargs()
+        kwargs['validation'] = Validation.objects.filter(pk=self.kwargs['pk']).last()
+        return kwargs
+
+    def form_valid(self, form):
+
+        user = self.request.user
+
+        validation = Validation.objects.filter(pk=self.kwargs['pk']).last()
+
+        if not validation:
+            raise PermissionDenied
+
+        if validation.validator != user:
+            raise PermissionDenied
+
+        if validation.status != 'PENDING_V':
+            raise PermissionDenied
+
+        # reject validation:
+        validation.comment_validator = form.cleaned_data['text']
+        validation.status = "REJECT"
+        validation.date_validation = datetime.now()
+        validation.save()
+
+        validation.content.sha_validation = None
+        validation.content.save()
+
+        # send PM
+        versioned = validation.content.load_version(sha=validation.version)
+        msg = render_to_string(
+            'tutorialv2/messages/validation_reject.md',
+            {
+                'content': versioned,
+                'url': versioned.get_absolute_url() + '?version=' + validation.version,
+                'validator': validation.validator,
+                'message_reject': '\n'.join(['> ' + a for a in form.cleaned_data['text'].split('\n')])
+            })
+
+        bot = get_object_or_404(User, username=settings.ZDS_APP['member']['bot_account'])
+        send_mp(
+            bot,
+            validation.content.authors.all(),
+            _(u"Rejet de la demande publication").format(),
+            validation.content.title,
+            msg,
+            True,
+            direct=False
+        )
+
+        messages.info(self.request, _(u'Le contenu a bien été refusé'))
+        self.success_url = reverse('validation:list')
+        return super(RejectValidation, self).form_valid(form)
+
+
+class AcceptValidation(LoginRequiredMixin, PermissionRequiredMixin, ModalFormView):
+    """Publish the content"""
+
+    permissions = ["tutorial.change_tutorial"]
+    form_class = AcceptValidationForm
+
+    modal_form = True
+
+    def get_form_kwargs(self):
+        kwargs = super(AcceptValidation, self).get_form_kwargs()
+        kwargs['validation'] = Validation.objects.filter(pk=self.kwargs['pk']).last()
+        return kwargs
+
+    def form_valid(self, form):
+
+        user = self.request.user
+        validation = Validation.objects.filter(pk=self.kwargs['pk']).last()
+
+        if not validation:
+            raise PermissionDenied
+
+        if validation.validator != user:
+            raise PermissionDenied
+
+        if validation.status != 'PENDING_V':
+            raise PermissionDenied
+
+        # get database representation and validated version
+        db_object = validation.content
+        versioned = db_object.load_version(sha=validation.version)
+        self.success_url = versioned.get_absolute_url(version=validation.version)
+
+        try:
+            published = publish_content(db_object, versioned, is_major_update=form.cleaned_data['is_major'])
+        except FailureDuringPublication as e:
+            messages.error(self.request, e.message)
+        else:
+            # save in database
+            is_update = db_object.sha_public and db_object.sha_public != ''
+            db_object.sha_public = validation.version
+            db_object.source = form.cleaned_data['source']
+            db_object.sha_validation = None
+
+            db_object.public_version = published
+
+            if form.cleaned_data['is_major']:
+                db_object.pubdate = datetime.now()
+
+            db_object.save()
+
+            # save validation object
+            validation.comment_validator = form.cleaned_data['text']
+            validation.status = "ACCEPT"
+            validation.date_validation = datetime.now()
+            validation.save()
+
+            # TODO: deal with other kind of publications (HTML, PDF, archive, ...)
+
+            if is_update:
+                msg = render_to_string(
+                    'tutorialv2/messages/validation_accept_update.md',
+                    {
+                        'content': versioned,
+                        'url': published.get_absolute_url_online(),
+                        'validator': validation.validator
+                    })
+            else:
+                msg = render_to_string(
+                    'tutorialv2/messages/validation_accept_content.md',
+                    {
+                        'content': versioned,
+                        'url': published.get_absolute_url_online(),
+                        'validator': validation.validator
+                    })
+
+            bot = get_object_or_404(User, username=settings.ZDS_APP['member']['bot_account'])
+            send_mp(
+                bot,
+                db_object.authors.all(),
+                _(u"Publication acceptée"),
+                versioned.title,
+                msg,
+                True,
+                direct=False
+            )
+
+            messages.success(self.request, _(u'Le contenu a bien été validé.'))
+            self.success_url = published.get_absolute_url_online()
+
+        return super(AcceptValidation, self).form_valid(form)
+
+
+class RevokeValidation(LoginRequiredMixin, PermissionRequiredMixin, SingleContentFormViewMixin):
+    """Unpublish a content and reverse the situation back to a pending validation"""
+
+    permissions = ["tutorial.change_tutorial"]
+    form_class = RevokeValidationForm
+    is_public = True
+
+    modal_form = True
+
+    def get_form_kwargs(self):
+        kwargs = super(RevokeValidation, self).get_form_kwargs()
+        kwargs['content'] = self.versioned_object
+        return kwargs
+
+    def form_valid(self, form):
+
+        versioned = self.versioned_object
+
+        if form.cleaned_data['version'] != self.object.sha_public:
+            raise PermissionDenied
+
+        validation = Validation.objects.filter(
+            content=self.object,
+            version=self.object.sha_public,
+            status='ACCEPT').prefetch_related('content__authors').last()
+
+        if not validation:
+            raise PermissionDenied
+
+        unpublish_content(self.object)
+
+        validation.status = "PENDING"
+        validation.validator = None  # remove previous validator
+        validation.date_validation = None
+        validation.save()
+
+        self.object.sha_public = None
+        self.object.sha_validation = validation.version
+        self.object.pubdate = None
+        self.object.save()
+
+        # send PM
+        msg = render_to_string(
+            'tutorialv2/messages/validation_revoke.md',
+            {
+                'content': versioned,
+                'url': versioned.get_absolute_url() + '?version=' + validation.version,
+                'admin': self.request.user,
+                'message_reject': '\n'.join(['> ' + a for a in form.cleaned_data['text'].split('\n')])
+            })
+
+        bot = get_object_or_404(User, username=settings.ZDS_APP['member']['bot_account'])
+        send_mp(
+            bot,
+            validation.content.authors.all(),
+            _(u"Dépublication"),
+            validation.content.title,
+            msg,
+            True,
+            direct=False
+        )
+
+        messages.success(self.request, _(u"Le tutoriel a bien été dépublié."))
+        self.success_url = self.versioned_object.get_absolute_url() + "?version=" + validation.version
+
+        return super(RevokeValidation, self).form_valid(form)
