+++ zds/utils/templatetags/target_tree.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 
 from zds.tutorialv2.utils import get_target_tagged_tree
-from zds.tutorialv2.models import Extract, Container
+from zds.tutorialv2.models.models_versioned import Extract, Container
 from django import template
